Reduce stale range from 365 days to 180 days
Summary:
We have a lot of stale issues that are receiving essentially no traffic. Previously we used to use `stalebot` which had an interval of 90 days but was somehow unreliable. Now we moved to the state GH actions but with an interval of 365 days which I believe is potentially too high.

Changelog:
[Internal] [Changed] - Reduce stale range from 365 days to 180 days

Created from CodeHub with https://fburl.com/edit-in-codehub

Reviewed By: GijsWeterings

Differential Revision: D34831199

fbshipit-source-id: 70f3f6cc97ac4add625061577e53cc80da987c7a